+++ package.json
@@ -41,8 +41,8 @@
     "request": "*",
     "serve-favicon": "~2.2.1",
     "socket.io": "*",
-    "promisified-request": "*",
-    "underscore": "*"
+    "underscore": "*",
+    "request-promise": "*"
   },
   "homepage": "http://www.getontrac.info",
   "devDependencies": {
