+++ server/routes/index.js
@@ -1,13 +1,15 @@
 var express = require('express');
 var router = express.Router();
 var moment = require('moment');
-var request = require("request");
+var request = require("request-promise");
 var cheerio = require("cheerio");
 var schedules = require("../modules/schedule");
 var achievements = require("../modules/achievements");
 var secrets = require("../secrets.json");
 
 var redir = "/";
+
+
 
 module.exports = function(io) {
   /* GET home page. */
@@ -16,350 +18,102 @@ module.exports = function(io) {
     console.log(req.trimester);
     if(req.loggedIn){
       req.toJade.title = "Your Home";
-
       res.render('home/homepage', req.toJade);
     }else{
       res.render('home/index', req.toJade);
     }
   });
 
+
+  // THE SP00KY LOGIN SYSTEM
   router.post('/login', function(req, res, next) {
-    var username = req.body.username;
-    var password = req.body.password;
+    var username = req.body.username.trim();
+    var password = req.body.password.trim();
+    if(!username || !password){
+      finalize("Please enter a username and password.");
+    }
     console.log("ATTEMPTING TO LOGIN AS "+username+": \n");
     var errs = [];
-    req.Student.findOne({username: username}).populate('courses', 'title mID').exec(function(err, user) {
-      if(err){errs.push('An error occured, please try again.'); done(); return;}
 
-      if(!user){
-        errs.push("0: Incorrect username or password."); // No user with that username found
-        done();
-      }else{
-        console.log("FOUND DATABASE MATCH");
-        // ATTEMPT INTRANET LOGIN
-        console.log(username + ":"+password);
-        var cookieJar = request.jar(); // To keep a persistent session
-        request({
-          url: 'https://intranet.regis.org/login/submit.cfm', //URL that the login form on the Intranet points to
-          jar: cookieJar,
-          method: 'POST',
-          //Lets post the following key/values as form
-          form: {
-            username: username,
-            password: password
-          }
-        }, function(error, response, body){
-            if(error) {
-                console.log(error);
-                errs.push("1: An error occured, please try again later.");
-                done(); return;
+    req.Student.findOne({username: username})
+      .populate('courses', 'title mID')
+      .exec(function(err, user) {
+        if(err){finalize("An error occured with the server. Please tell Frank!");}
+        if(!user){finalize("0: Incorrect username or password.");}else{
+          // The username passed exists
+          var cookieJar = request.jar(); // For the session
+          request({
+            url: 'https://intranet.regis.org/login/submit.cfm', //URL that the login form on the Intranet points to
+            jar: cookieJar,
+            method: 'POST',
+            //Lets post the following key/values as form
+            form: {
+              username: username,
+              password: password
+            },
+            resolveWithFullResponse: true,
+            simple: false
+          }).then(function(response){
+            var body = response.body;
+            if(body.indexOf("url=https://www.regis.org/login.cfm?Failed=1") > -1){
+              finalize("2: Incorrect username or password.");
             }else{
-              console.log("1: "+response.statusCode);
-              var $ = cheerio.load(body);
-              var title = $("title").text();
-              //console.log(body);
-              if(body.indexOf("url=https://www.regis.org/login.cfm?Failed=1") > -1){ // This is the HTML rendered when you pass Incorrect credentials to the Intranet login
-                errs.push("2: Incorrect username or password.");
-                done();
-              }else{
-                console.log("LOGGED IN AS "+user.username+" SUCCESSFULLY.")
-                request({
-                    url: 'http://intranet.regis.org/functions/view_profile.cfm', //URL to hit
-                    jar: cookieJar,
-                    method: 'GET'
-                }, function(error, response, html){
-                  if(error) {
-                      console.log(error);
-                      errs.push("1: Failed to register. Please try again later.");
-                      done(); return;
-                  }else{
-                    var $ = cheerio.load(html);
+              //LOGGED IN TO INTRANET SUCCESSFULLY (THIS IS A REAL USER)
+              if(user.registered == false || user.trimester_updated_in != req.trimester){
+                // NOT REGISTERED, get his Student ID and register him!
+                if(user.registered == false){
+                  console.log("FIRST LOGIN FOR "+username);
+                  user.login_count = 0;
+                  user.last_point_login_time = new Date();
+                  user.registered_date = new Date();
+                  user.registered = true;
+                  user.nickname = user.firstName;
 
-                    // SUCCESSFULLY LOGGED IN, IS THIS USER REGISTERED?
-
-                    if(user.registered == false || user.trimester_updated_in != req.trimester){
-
-                      // NO, get his Student ID and register him!
-                      if(user.registered == false){
-                        console.log("FIRST LOGIN FOR "+username);
-                        var id = $("td:contains('Locker Number:')").next().text().trim();
-                        //console.log(id);
-                        user.locker_number = id;
-                        user.login_count = 0;
-                        user.last_point_login_time = new Date();
-                        user.registered_date = new Date();
-                        user.registered = true;
-                        user.nickname = user.firstName;
-                      }else{
-                        console.log("UPDATING INFO FOR "+username);
-                      }
-
-                      if(user.username == "fmatranga18")
-                        user.rank = 7;
-
-                      // THE GENIUS PART
-                      // Once logged in to a brand new user, download his schedule from the Intranet and parse it
-                      request({
-                          url: 'http://intranet.regis.org/downloads/outlook_calendar_import/outlook_schedule_download.cfm', //URL to hit
-                          jar: cookieJar,
-                          method: 'GET'
-                      }, function(error, response, text){
-                        if(error){
-                          console.log(error);
-                          errs.push("1: Failed to register. Please try again later.");
-                          done(); return;
-                        }
-                        console.log("DOWNLOADED SCHEDULE FOR "+username);
-                        var scheduleLines = text.match(/[^\r\n]+/g);
-                        var good = [];
-                        var schedule = {
-                          scheduleDays: {},
-                          dayClasses: {}
-                          /*
-                          scheduleDays: {
-                            '09/22/15': 'E'
-                          },
-                          dayClasses: {
-                            'A': [A DAY CLASSES],
-                            etc...
-                          }
-                          */
-                        };
-
-                        scheduleLines.forEach(function(line) {
-                          good.push(line.split("\t"));
-                        });
-                        var actualClasses = [];
-                        good.forEach(function(line) {
-                          var dateS = line[0]; // String date in format MM/DD/YY
-                          if(moment(dateS, "MM/DD/YY").isValid()){
-                            if(line[4].indexOf(" Day") > -1){ // This line is stating a Schedule Day
-                              schedule.scheduleDays[dateS] = line[4].replace(" Day", "");
-                            }else{
-                              actualClasses.push(line);
-                            }
-                          }
-                        });
-
-                        var dIndex = 0;
-                        if(req.trimester == 2)
-                          dIndex = Math.floor(actualClasses.length / 2)
-                        else if(req.trimester == 3)
-                          dIndex = actualClasses.length-70
-
-                        actualClasses = actualClasses.slice(dIndex, dIndex+60);
-                        //console.log(actualClasses);
-
-                        var doneF = false;
-                        var i = 0;
-                        while(doneF == false && i < actualClasses.length){
-                          var line = actualClasses[i];
-                          var dateS = line[0];
-                          var sd = schedule.scheduleDays[dateS];
-
-                          if(!schedule.dayClasses[sd])
-                            schedule.dayClasses[sd] = [];
-
-                          var room = (isNaN(line[5]) ? line[5] : "Room "+line[5]);
-
-                          var duration = moment(line[3], "hh:mm A").diff(moment(line[1], "hh:mm A"), 'minutes');
-                          var period = {
-                            //date: moment(line[0], "MM/DD/YY").toDate(),
-                            startTime: line[1],
-                            endTime: line[3],
-                            duration: duration,
-                            className: line[4],
-                            room: room
-                          };
-
-                          var length = schedule.dayClasses[sd].filter(function(p){
-                            if(p.className == line[4])
-                              return true;
-                            return false;
-                          }).length;
-                          if(length == 0)
-                            schedule.dayClasses[sd].push(period);
-
-                          i ++;
-                        }
-
-                        user.scheduleObject = schedule;
-                        ['A', 'B', 'C', 'D', 'E'].forEach(function(sd){
-                          var filledPeriods = [];
-                          console.log(sd);
-                          // AM ADVISEMENT
-                          var amAdv = {
-                            startTime: '08:40 AM',
-                            endTime: '08:50 AM',
-                            className: 'Morning Advisement',
-                            duration: 10,
-                            room: 'Homeroom'
-                          };
-                          filledPeriods.push(amAdv);
-
-                          var periods = schedule.dayClasses[sd];
-                          var lastPeriod = amAdv;
-
-                          periods.forEach(function(period) {
-                            if(lastPeriod.endTime !== period.startTime){
-                              var room = "Anywhere";
-                              var cName = "Unstructured Time";
-                              var grade = user.grade;
-                              var lunch = moment(lastPeriod.endTime, "hh:mm A").startOf('day').hour(11).minute(30);
-                              switch(grade){
-                                case 9:
-                                case 10:
-                                  lunch = moment(lastPeriod.endTime, "hh:mm A").startOf('day').hour(11).minute(30);
-                                  break;
-                                case 11:
-                                case 12:
-                                  lunch = moment(lastPeriod.endTime, "hh:mm A").startOf('day').hour(12).minute(10);
-                                  break;
-                              }
-
-
-                              console.log(period.duration);
-
-                              if(moment(lastPeriod.endTime, "hh:mm A").isSame(lunch)){
-                                filledPeriods.push({
-                                  room: "Cafeteria",
-                                  startTime: lunch.format("hh:mm A"),
-                                  endTime: moment(lunch).add(40, 'minutes').format("hh:mm A"),
-                                  duration: 40,
-                                  className: "Lunch"
-                                });
-                              }else{
-                                filledPeriods.push({
-                                  room: "Anywhere",
-                                  startTime: lastPeriod.endTime,
-                                  endTime: period.startTime,
-                                  duration: moment(lastPeriod.endTime, "hh:mm A").diff(moment(period.startTime, "hh:mm A"), 'minutes'),
-                                  className: "Unstructured Time"
-                                });
-                              }
-                            }
-                            //console.log(user.courses);
-                            var myC = user.courses;
-                            myC.forEach(function(course) {
-                              if(period.className.trim().indexOf(course.title.trim()) > -1 || course.title.trim() == period.className.trim()){
-                                console.log("MATCHED");
-                                period.mID = course.mID;
-                              }else{
-                                //console.log("COULDN'T MATCH "+period.className + " TO COURSE");
-                              }
-                            });
-
-                            lastPeriod = period;
-                            filledPeriods.push(period);
-                          });
-
-                          // End of day free
-                          if(filledPeriods[filledPeriods.length-1].endTime !== "02:50 PM"){
-                            var duration = moment(lastPeriod.endTime, "hh:mm A").hour(14).minute(50).diff(moment(lastPeriod.endTime, "hh:mm A"), 'minutes');
-                            filledPeriods.push({
-                              room: "Anywhere",
-                              startTime: lastPeriod.endTime,
-                              endTime: "02:50 PM",
-                              duration: duration,
-                              className: "Unstructured Time"
-                            });
-                          }
-
-                          // PM ADVISEMENT
-                          filledPeriods.push({
-                            startTime: '02:50 PM',
-                            endTime: '02:55 PM',
-                            duration: 20,
-                            className: 'Afternoon Advisement',
-                            room: 'Homeroom'
-                          });
-
-                          schedule.dayClasses[sd] = filledPeriods;
-                        });
-
-                        console.log("PARSED SCHEDULE FOR "+username);
-                        user.scheduleObject = schedule;
-                        user.trimester_updated_in = req.trimester;
-                        done();
-                      });
-
-                      new req.Log({who: user.username, what: "Registration."}).save();
-                      if(req.toJade.production){
-                        require("../modules/mailer")(user.email, "Welcome to OnTrac!",
-                          "Hello, "+user.firstName+". Thank you for using OnTrac. It \
-                          is currently in Alpha, meaning it has not fully been released \
-                          yet and is in active development. Thank you for being an Alpha Tester!");
-                      }
-                    }else{
-                      done();
-                    }
-                  }
+                  if(user.username == "fmatranga18")
+                    user.rank = 7;
+                }else{
+                  console.log("UPDATING INFO FOR "+username);
+                }
+                console.log("DOWNLOADING CLASS SCHEDULE FOR "+username);
+                return request({
+                  url: 'http://intranet.regis.org/downloads/outlook_calendar_import/outlook_schedule_download.cfm', //URL to hit
+                  jar: cookieJar,
+                  method: 'GET',
+                  resolveWithFullResponse: true,
+                  simple: false
                 });
+              }else{
+                finalize(null);
               }
             }
-        });
-      }
-      function done(){
-        if(errs.length > 0){
-          res.json({errors: errs});
-        }else{
-
-          var uA = user.achievements;
-          achievements.forEach(function(ach){
-            if(uA.indexOf(ach.id) == -1){
-              if(ach.check(user)){
-                user.achievements.push(ach.id);
-                user.points += ach.reward;
-                req.session.info.push("You have been awarded "+ach.reward+" points for achieving '"+ach.name+"'!");
-              }
-            }
-          });
-
-          user.login_count +=1;
-
-          // Give points if last time points were given was over 5 minutes ago
-          var fiveMinAgo = moment().subtract(5, 'minutes');
-          if(moment(user.last_point_login_time).isBefore(fiveMinAgo)){
-            user.points += 10;
-            req.session.info.push("You have been rewarded 10 points for logging in.");
-            user.last_point_login_time = new Date();
-          }
-          req.user = user;
-          req.session.currentUser = user;
-          req.session.currentUser.last_login_time = new Date();
-          user.save(function(err) {
-            if(err)console.error(err);
-          });
-
-          var sd = user.scheduleObject.scheduleDays[moment().format("MM/DD/YY")];
-          if(sd){
-            //console.log(user.scheduleObject.dayClasses);
-            req.session.todaysInfo = {scheduleDay: sd, periods: user.scheduleObject.dayClasses[sd]};
-            //console.log(req.session.todaysInfo);
-          }
-
-          new req.Log({who: user.username, what: "Login."}).save();
-          req.session.quietlogin = false; // The actual user logged in, not an admin
-
-          io.sockets.emit('new-login', {username: username});
-          io.sockets.emit('message', {username: "server", message: username+" has logged in.", when: Date.now()});
-          res.json({success: true});
-          console.log("COMPLETED LOGIN");
+          }).then(function(response) {
+            var body = response.body;
+            schedules.generateSchedule(body);
+          })
+          .catch(finalize);
         }
+      });
 
+
+    function finalize(err){
+      console.log(err);
+      if(err !== null){
+        err = err.message || err;
+        res.json({errors: [err]});
+        return;
       }
-    });
+
+      return;
+    }
   });
 
   router.get('/logout', function(req, res) {
     if(req.loggedIn){
-
       if(req.session.quietlogin == false){
         new req.Log({who: req.currentUser.username, what: "Logout."}).save();
         io.sockets.emit('new-logout', {username: req.currentUser.username});
         io.sockets.emit('message', {username: "server", message: req.currentUser.username+" has logged out.", when: Date.now()});
       }
-
       delete req.session.currentUser;
       delete req.currentUser;
       delete req.loggedIn;
